backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torch-geometric==2.3.1
rdkit-pypi==2023.9.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
pydantic==2.5.0 